Move Surgury checks over to Shared, remove from server, add proper tooltips, etc
