+++ Content.Client/_Starlight/Medical/Surgery/SurgeryBui.cs
@@ -375,6 +375,9 @@ public sealed class SurgeryBui : BoundUserInterface
                         case StepInvalidReason.NotEnoughReagent:
                             stepName.AddMarkupOrThrow(" [color=red](Missing Reagent)[/color]");
                             break;
+                        case StepInvalidReason.MissingLimb:
+                            stepName.AddMarkupOrThrow(" [color=red](Can't attach as limb)[/color]");
+                            break;
                     }
                 }
                 else
