+++ Content.Shared/_Starlight/Medical/Surgery/StepInvalidReason.cs
@@ -10,4 +10,5 @@ public enum StepInvalidReason
     DisabledTool,
     TooHigh,
     NotEnoughReagent,
+    MissingLimb,
 }
